% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFI_soil_codes.R
\name{ANFI_soil_codes}
\alias{ANFI_soil_codes}
\title{Soil groups of the Austrian National Forest Inventory (Forstliche Bundesversuchsanstalt 1981) transcribed to the FAO-UNESCO soil groups}
\usage{
ANFI_soil_codes(Code = 1:26)
}
\arguments{
\item{Code}{1-26. Default : 1:26.}
}
\value{
A tibble with the selected code(s).
}
\description{
From transcription by Sonja Vospernik. \url{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}
}
\examples{
ANFI_soil_codes()
}
