% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFI_vegetation_codes.R
\name{ANFI_vegetation_codes}
\alias{ANFI_vegetation_codes}
\title{Vegetation types of ANFI (1981), from Monserud and Sterba (1996)}
\usage{
ANFI_vegetation_codes(Code = 1:20)
}
\arguments{
\item{Code}{1-20. Default (1:20)}
}
\value{
Tibble.
}
\description{
Vegetation types of ANFI (1981), from Monserud and Sterba (1996)
}
\examples{
ANFI_vegetation_codes()
}
