% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFI_growth_zone.R
\name{Austria_growth_zone}
\alias{Austria_growth_zone}
\title{Growth Zones for Austria (ANFI)}
\usage{
Austria_growth_zone(Code = 1:21)
}
\arguments{
\item{Code}{1:21.}
}
\value{
Tibble.
}
\description{
Growth Zones for Austria (ANFI)
}
