% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Hasenauer_1997_crown_diameter_open_grown_Austria.R
\name{Hasenauer_1997_crown_diameter_open_grown_Austria}
\alias{Hasenauer_1997_crown_diameter_open_grown_Austria}
\title{Calculate the Crown Diameter of Open grown trees in Austria following Hasenauer 1997.}
\source{
Hasenauer, H. 1997. Dimensional relationships of open-grown trees in Austria. Forest Ecology & Management. Vol 96:3. pp. 197-206. DOI: \url{https://doi.org/10.1016/S0378-1127(97)00057-1}
}
\usage{
Hasenauer_1997_crown_diameter_open_grown_Austria(
  species,
  diameter_cm,
  log_correction = FALSE
)
}
\arguments{
\item{species}{One of: "Picea abies", "Abies alba", "Larix decidua","Pinus sylvestris","Pinus nigra","Pinus cembra","Fagus sylvatica","Quercus","Fraxinus","Acer", or another deciduous tree species.}

\item{diameter_cm}{Diameter of the tree at breast height, cm.}

\item{log_correction}{TRUE or FALSE (Default) to apply logarithmic correction.}
}
\value{
Crown diameter.
}
\description{
Calculate the Crown Diameter of Open grown trees in Austria following Hasenauer 1997.
}
