% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_zone.R
\name{influence_zone}
\alias{influence_zone}
\title{Open-grown tree crown area}
\usage{
influence_zone(crown_diameter, nrep = 1)
}
\arguments{
\item{crown_diameter}{Diameter of an open-grown tree in meter}

\item{nrep}{represented stem number, default=1, data is not from
a sample, i.e. each tree represents 1 tree per hectare}
}
\description{
Calculate the represented open-grown tree crown area
in m²
}
\details{
Calculates the represented open-grown crown area from an open-grown tree
diameter, e.g. [Prognaus::Hasenauer_1997_crown_diameter_open_grown_Austria()]
assuming circular crown areas and multiplies
them by the represented stem number (e.g. in fixed area plots or
angle count samples)
}
