% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Krajicek_1961_crown_competition_factor_PROGNAUS.R
\name{Krajicek_1961_crown_competition_factor_PROGNAUS}
\alias{Krajicek_1961_crown_competition_factor_PROGNAUS}
\title{Krajicek's Crown competition factor (1961) as implemented in PROGNAUS}
\source{
Krajicek, J. E., Brinkmann, K. E., and Gingrich, S. F. (1961): Crown Competition - a Measure of Density. For. Sci. 7, 35-42.
}
\usage{
Krajicek_1961_crown_competition_factor_PROGNAUS(
  species,
  diameter_cm,
  nrep = 1,
  log_correction = 0
)
}
\arguments{
\item{species}{One of: "Picea abies", "Abies alba", "Larix decidua","Pinus sylvestris","Pinus nigra","Pinus cembra","Fagus sylvatica","Quercus","Fraxinus","Acer", or another deciduous tree species.}

\item{diameter_cm}{Diameter of the tree at breast height, cm.}

\item{nrep}{represented stem number, default=1, data is not from
a sample, i.e. each tree represents 1 tree per hectare}

\item{log_correction}{TRUE or FALSE (Default) to apply logarithmic correction.}

\item{area}{Area of plot/stand in m^2. Default = 100 m^2.}
}
\description{
Calculates the crown competition factor
}
\details{
Calculates the crown competition factor according to Krajicek et al. (1961)
using the open-grown tree diameter functions developed
by Hasenauer (1997) [Prognaus::Hasenauer_1997_crown_diameter_open_grown_Austria()].

Circular crown areas are assumed
and multiplied by the represented stem number (e.g. in fixed area plots or
angle count samples), e.g. [Prognaus::influence_zone()]
}
