% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monserud_1996_BA_increment_Austria.R
\name{Monserud_1996_BA_increment_Austria}
\alias{Monserud_1996_BA_increment_Austria}
\title{Basal area increment model Prognaus}
\source{
Hasenauer, H. (2000): Die simultanen Eigenschaften von Waldwachstumsmodellen. Paul Parey Verlag. Berlin, 131 pp.

Monserud, R. A. and Sterba, H. (1996): A Basal Area Increment Model for Individual Trees Growing in Even- and Unevenaged Forest Stands in Austria. For. Ecol. and Manage. 80, 57-80.
}
\usage{
Monserud_1996_BA_increment_Austria(
  species = "Picea abies",
  diameter_cm,
  crown_ratio,
  BAL,
  crown_competition_factor,
  altitude = 600,
  incline_percent = 50,
  aspect_degree = 180,
  Humus_F_cm = 3,
  Humus_H_cm = 2,
  soil_depth_cm = 50,
  soil_moisture = 3,
  slope_position = 2,
  soil_type = 9,
  vegetation = 1,
  ANFI_region = 13,
  log_correction = 0,
  period = 5
)
}
\arguments{
\item{species}{One of.. : "Picea abies","Abies alba","Larix decidua","Pinus sylvestris","Pinus nigra","Pinus cembra", "Fagus sylvatica","Quercus", or other deciduous tree.}

\item{diameter_cm}{Diameter at breast height of tree.}

\item{crown_ratio}{Crown ratio of tree, e.g. [Prognaus::Hasenauer_1996_crown_ratio_Austria()]}

\item{BAL}{Basal area of larger trees (Wykoff, 1990), e.g. [Prognaus::Wykoff_1990_BAL()]}

\item{crown_competition_factor}{Crown competition factor (Krajicek, 1961), e.g.[Prognaus::Krajicek_1961_crown_competition_factor_PROGNAUS()]}

\item{altitude}{Elevation in m (Range: 200-2200 m)}

\item{incline_percent}{Slope in percent}

\item{aspect_degree}{Aspect in degree}

\item{Humus_F_cm}{Thickness of F-humus layer in cm}

\item{Humus_H_cm}{Thickness of H-humus layer in cm}

\item{soil_depth_cm}{Soil depth in cm, model uses dummy variable which is 1 for soil depth <=30cm}

\item{soil_moisture}{Soil moisture,Type 1="Dry",2="Mesic",3="Mesic-moist",4="Moist",5="Wet"}

\item{slope_position}{ANFI Slope position. Type 1="Upper slope,crest",2="Middle slope",3="Lower slope",4="Toe slope",5="Valley plain",6="Plain",7="Depression".}

\item{soil_type}{Soil type, code according to [Prognaus::ANFI_soil_codes()]}

\item{vegetation}{Vegetation type, code according to [Prognaus::ANFI_vegetation_codes()]}

\item{ANFI_region}{Growth district, code according to [Prognaus::Austria_growth_zone()]}

\item{log_correction}{Dummy variable, default=0, if 1 half of the model MSE is
added to correct for logarithmic transformation bias}

\item{period}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and crown_competition_factor}
}
\description{
Calculates the basal area increment in cm²/period
}
\details{
Calculates the basal area increment of Prognaus. The model is based
on Monserud and Sterba (1996); Coefficients are taken from Hasenauer (2000).
}
\examples{
Monserud_1996_BA_increment_Austria(species="Picea abies",diameter_cm=30,crown_ratio = 0.5,BAL=0,crown_competition_factor=200)
}
