% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Nachtmann_2005_height_increment_Prognaus_Austria.R
\name{Nachtmann_2005_height_increment_Prognaus_Austria}
\alias{Nachtmann_2005_height_increment_Prognaus_Austria}
\title{Calculate the height increment for trees as in Prognaus.}
\source{
Nachtmann, G. (2005): Der Höhenzuwachs österreichischer Baumarten in Abhängigkeit von Standort und Konkurrenz. Diploma thesis. Boku, University of Natural Resources and Life Sciences. 65 pp.
}
\usage{
Nachtmann_2005_height_increment_Prognaus_Austria(
  species = "Picea abies",
  height_m,
  dominant_height_m,
  BAL,
  crown_competition_factor,
  altitude = 600,
  incline_percent = 50,
  aspect_degree = 180,
  soil_depth_cm = 50,
  soil_moisture = 3,
  slope_position = 2,
  soil_type = 9,
  ANFI_region = 13,
  period = 5
)
}
\arguments{
\item{species}{One of: "Picea abies", "Abies alba", "Larix decidua","Pinus sylvestris","Pinus nigra","Pinus cembra","Fagus sylvatica","Quercus","Fraxinus","Acer".}

\item{height_m}{Tree height in m}

\item{dominant_height_m}{Dominant height in m}

\item{BAL}{Basal area of larger trees (Wykoff, 1990), e.g. [Prognaus::Wykoff_1990_BAL()]}

\item{crown_competition_factor}{Crown competition factor (Krajicek, 1961), e.g. [Prognaus::crown_competition_factor()]}

\item{altitude}{Elevation in m (Range: 200-2200 m)}

\item{incline_percent}{Slope in percent}

\item{aspect_degree}{Aspect in degree}

\item{soil_depth_cm}{Soil depth in cm, model uses dummy variable which is 1 for soildepth <=30cm}

\item{soil_moisture}{,Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{slope_position}{ANFI Slope position. Type 1="Upper slope,crest",2="Middle slope",3="Lower slope",4="Toe slope",5="Valley plain",6="Plain",7="Depression".}

\item{soil_type}{Soil type e.g. [Prognaus::ANFI_soil_codes()]}

\item{ANFI_region}{Growth district, e.g. [Prognaus::ANFI_growth_zone()]}

\item{period}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and CCF}
}
\description{
Calculates the height increment in m/period as in Prognaus.
The model is an Evolon-model, c.f. Mende and Albrecht (2001).
Coding of tree species, soil moisture, relief, soil type, vegetation type
and growth district uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
Nachtmann_2005_height_increment_Prognaus_Austria(species="Picea abies",height_m=30,dominant_height_m=30,BAL=0,crown_competition_factor=200)

height2012<-30
height2015<-height2012+Nachtmann_2005_height_increment_Prognaus_Austria(species="Picea abies",height_m=30,dominant_height_m=30,BAL=0,crown_competition_factor=200,period=3)
}
