% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pollanschutz_1974_get_form_quotient_Austria.R
\name{Pollanschutz_1974_get_form_quotient_Austria}
\alias{Pollanschutz_1974_get_form_quotient_Austria}
\title{Form factors for different tree species as used in Prognaus.}
\source{
Pollanschütz, J. (1974): Formzahlfunktionen der Hauptbaumarten
Österreichs. AFZ 85, 341- 343.
}
\usage{
Pollanschutz_1974_get_form_quotient_Austria(species, diameter_cm, height_m)
}
\arguments{
\item{species}{One of: "Picea abies", "Abies alba", "Larix decidua",
"Pinus sylvestris","Pinus nigra","Pinus cembra","Fagus sylvatica","Quercus",
"Fraxinus","Acer", or another tree species.}

\item{diameter_cm}{Diameter of tree at breast height.}

\item{height_m}{Tree height in metres.}
}
\description{
Calculates the form factor;
form factors are dimensionless. For use together with tree_volume_Prognaus().
}
\examples{
#Volume of an individual tree in m3, see also vPoll
(dbh/100)^2*pi/4*height*Pollanschutz_1974_get_form_quotient_Austria(species="Picea abies",diameter_cm=30,height_m=30)
}
