% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vospernik_diameter_increment_PROGNAUS.R
\name{Vospernik_diameter_increment_PROGNAUS}
\alias{Vospernik_diameter_increment_PROGNAUS}
\title{Diameter increment model Prognaus}
\usage{
Vospernik_diameter_increment_PROGNAUS(
  species = "Picea abies",
  diameter_cm,
  crown_ratio,
  BAL,
  crown_competition_factor,
  altitude = 600,
  incline_percent = 50,
  aspect_degree = 180,
  Humus_F_cm = 3,
  Humus_H_cm = 2,
  soil_depth_cm = 50,
  soil_moisture = 3,
  slope_position = 2,
  soil_type = 9,
  vegetation = 1,
  ANFI_region = 13,
  log_correction = 0,
  period = 5
)
}
\arguments{
\item{species}{One of.. : "Picea abies","Abies alba","Larix decidua","Pinus sylvestris","Pinus nigra","Pinus cembra", "Fagus sylvatica","Quercus", or other deciduous tree.}

\item{diameter_cm}{Diameter at breast height of tree.}

\item{crown_ratio}{Crown ratio of tree, e.g. [Prognaus::Hasenauer_1996_crown_ratio_Austria()]}

\item{BAL}{Basal area of larger trees (Wykoff, 1990), e.g. [Prognaus::Wykoff_1990_BAL()]}

\item{crown_competition_factor}{Crown competition factor (Krajicek, 1961), e.g.[Prognaus::Krajicek_1961_crown_competition_factor_PROGNAUS()]}

\item{altitude}{Elevation in m (Range: 200-2200 m)}

\item{incline_percent}{Slope in percent}

\item{aspect_degree}{Aspect in degree}

\item{Humus_F_cm}{Thickness of F-humus layer in cm}

\item{Humus_H_cm}{Thickness of H-humus layer in cm}

\item{soil_depth_cm}{Soil depth in cm, model uses dummy variable which is 1 for soil depth <=30cm}

\item{soil_moisture}{Soil moisture,Type 1="Dry",2="Mesic",3="Mesic-moist",4="Moist",5="Wet"}

\item{slope_position}{ANFI Slope position. Type 1="Upper slope,crest",2="Middle slope",3="Lower slope",4="Toe slope",5="Valley plain",6="Plain",7="Depression".}

\item{soil_type}{Soil type, code according to [Prognaus::ANFI_soil_codes()]}

\item{vegetation}{Vegetation type, code according to [Prognaus::ANFI_vegetation_codes()]}

\item{ANFI_region}{Growth district, code according to [Prognaus::Austria_growth_zone()]}

\item{log_correction}{Dummy variable, default=0, if 1 half of the model MSE is
added to correct for logarithmic transformation bias}

\item{period}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and crown_competition_factor}
}
\description{
Helper function to [Prognaus::Monserud_1996_BA_increment_Austria()].
Transforms output to the diameter increment in cm/period.
}
\details{
Calculates the diameter increment of Prognaus. The model calls the
basal area increment function of Prognaus, e.g. [Prognaus::Monserud_1996_BA_increment_Austria()] and converts it to diameter
increment.
}
\examples{
Vospernik_diameter_increment_PROGNAUS(species="Picea abies",diameter_cm=30,crown_ratio = 0.5,BAL=0,crown_competition_factor=200)

dbh2012<-30
dbh2015<-dbh2012+Vospernik_diameter_increment_PROGNAUS(species="Picea abies",diameter_cm=30,crown_ratio = 0.5,BAL=0,crown_competition_factor=200,period=3)
}
