% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wykoff_1990_BAL.R
\name{Wykoff_1990_BAL}
\alias{Wykoff_1990_BAL}
\title{Basal area of larger trees}
\usage{
Wykoff_1990_BAL(diameter_cm, basal_area_m2_ha)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height in cm.}

\item{basal_area_m2_ha}{Basal area of tree, m2/ha. (Or the BA which it represents).}
}
\value{
A vector of the Basal area of the larger trees for each tree.
}
\description{
Calculates the basal area of larger trees according to Wykoff 1990.
}
